ft: relocate client server version to constants
