+++ src/prefect/client/orchestration.py
@@ -31,6 +31,7 @@ import prefect
 import prefect.exceptions
 import prefect.settings
 import prefect.states
+from prefect.client.constants import SERVER_API_VERSION
 from prefect.client.schemas import FlowRun, OrchestrationResult, TaskRun
 from prefect.client.schemas.actions import (
     ArtifactCreate,
@@ -202,13 +203,6 @@ class PrefectClient:
             httpx_settings.setdefault("verify", False)
 
         if api_version is None:
-            # deferred import to avoid importing the entire server unless needed
-            try:
-                from prefect.server.api.server import SERVER_API_VERSION
-            except ImportError:
-                # support the case where server components are not available
-                SERVER_API_VERSION = "0.8.4"
-
             api_version = SERVER_API_VERSION
         httpx_settings["headers"].setdefault("X-PREFECT-API-VERSION", api_version)
         if api_key:
